Added event_name right-hand label, label now shows in browsers
From https://docs.moodle.org/dev/lib/formslib.php_Form_Definition#checkbox: "The third parameter for this element is the label to display on the left side of the form. You can also supply a string as a fourth parameter to specify a label that will appear on the right of the element.".